Use mps if it's available on Apple Silicon devices.
- Use pytorch 2.5.0 instead of nightly.

- FIX: activation error on MPS

  MPS can't silu activation and creates randomly broken results
  if tensor memory format is not contiguous.

  This is not happening on macOS 15 and later because it's using
  native stride but macOS 14 is affected.

+++ pyramid_dit/mmdit_modules/modeling_pyramid_mmdit.py
@@ -28,7 +28,7 @@ from IPython import embed
 def rope(pos: torch.Tensor, dim: int, theta: int) -> torch.Tensor:
     assert dim % 2 == 0, "The dimension must be even."
 
-    scale = torch.arange(0, dim, 2, dtype=torch.float64, device=pos.device) / dim
+    scale = torch.arange(0, dim, 2, dtype=pos.dtype, device=pos.device) / dim
     omega = 1.0 / (theta**scale)
 
     batch_size, seq_length = pos.shape
